+++ client/web/src/tracking/eventLogger.ts
@@ -2,6 +2,7 @@ import cookies, { CookieAttributes } from 'js-cookie'
 import * as uuid from 'uuid'
 
 import { TelemetryService } from '@sourcegraph/shared/src/telemetry/telemetryService'
+import { UTMMarker } from '@sourcegraph/shared/src/tracking/utm'
 
 import { browserExtensionMessageReceived } from './analyticsUtils'
 import { serverAdmin } from './services/serverAdminWrapper'
@@ -244,22 +245,16 @@ function handleQueryEvents(url: string): void {
     stripURLParameters(url, ['utm_campaign', 'utm_source', 'utm_medium', 'badge'])
 }
 
-interface EventQueryParameters {
-    utm_campaign?: string
-    utm_source?: string
-    utm_medium?: string
-}
-
 /**
  * Get pageview-specific event properties from URL query string parameters
  */
-function pageViewQueryParameters(url: string): EventQueryParameters {
+function pageViewQueryParameters(url: string): UTMMarker {
     const parsedUrl = new URL(url)
 
     const utmSource = parsedUrl.searchParams.get('utm_source')
     const utmCampaign = parsedUrl.searchParams.get('utm_campaign')
 
-    const utmProps = {
+    const utmProps: UTMMarker = {
         utm_campaign: utmCampaign || undefined,
         utm_source: utmSource || undefined,
         utm_medium: parsedUrl.searchParams.get('utm_medium') || undefined,
@@ -275,6 +270,17 @@ function pageViewQueryParameters(url: string): EventQueryParameters {
         eventLogger.log('CodeMonitorEmailLinkClicked')
     } else if (utmSource === 'hubspot' && utmCampaign?.match(/^cloud-onboarding-email(.*)$/)) {
         eventLogger.log('UTMCampaignLinkClicked', utmProps, utmProps)
+    } else if (
+        [
+            'safari-extension',
+            'firefox-extension',
+            'chrome-extension',
+            'phabricator-integration',
+            'bitbucket-integration',
+            'gitlab-integration',
+        ].includes(utmSource ?? '')
+    ) {
+        eventLogger.log('UTMCodeHostIntegration', utmProps, utmProps)
     }
 
     return utmProps